+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## Next
 
+## 🐏 Release 66.3.0 - 2024-11-18
+
 ### Added
 * Skiplink: New component ([#2746](https://github.com/dso-toolkit/dso-toolkit/issues/2746))